Introduce new weight models
WeightUnit will be used by our persistence layer in order to let users track
movement weights with a specific unit. Storing weights with a specific unit
type will help prevent dataloss when converting to/from display and storage
weights. For example, if we stored things in kilos we'd end up storing 225
pounds as 102.05840462301894 pounds. Besides being overly specific, it's
reasonable to assume that using this number in math operations could lead to
dropping some of those decimal points and change the value that the user
actually cares about.