Experiment with multi-field forms
I don't know what you'd call this, but we're going to need to let users add
arbitrary numbers of sets. This means we'll need to let users add any number
of fields to the form to represent their sets. I didn't know how to make this
work, so I wanted to do a small experiment to figure out how this worked.

Turns out, while it's a bit unintuitive at first, the end result isn't that
bad. The gist is that Yesod is smarter than it would seem on the surface here,
and will concatenate all fields with the same name into a list of values for
you automatically. All that's left is to write a small custom field that
understands that there might be many of them.

This commit implements one such field, and actually includes the code for
adding new fields to the form for the sake of simplicity. There's very little
data validation here, we're basically just saying "give me all of the data and
I'll deal with it", but it gets us to where we want to be: a single field
(technically a fieldset, I suppose) that is able to dynamically add arbitrary
numbers of values to itself.