Experiment with Elm for implementing workout form
Trying to wrap my head around how this (fairly complex) form would work in
Yesod made me start to wonder if something like Elm wouldn't be better suited
for the job. This is a small experiment in implementing the most basic
component I'll need for inputting workout info.

I think this experiment was a huge success, and I think that I'll probably
move forward with this model. Parsing this info at the client level and then
simply sending the results to Yesod will also let me focus on an API for the
backend, which will be necessary once I build out the mobile client anyway.

This component has 3 fields:

1. A reps field. This field can take reps for compound movements by entering
   the reps separated by "+" (and any amount of whitespace).
2. The weight for the movements. This field expects an Integer
3. A dropdown for selecting the unit of measurement. This is limited to Kilos
   and Pounds, and so crashes in the event that it gets a unit of measurement
   it doesn't expect.

The result is parsed and formatted for display to the right of the fields for
the sake of demonstration.

Some improvements I could make:

- Limit input for reps to numbers, whitespace, and "+"
- Limit input for weight to numbers

Next steps:

- Build a component that can dynamically add/remove these components and
  combine the result into something more useful.
- Add data validation to ensure that each list of reps is the same length as
  the rest and also matches the list of movements in the parent.