Remove main function from SingleSet
We don't need this to be a standalone component anymore, so we don't need this
main function.

We'll also go ahead and add the required exports for using this component from
a parent, and re-add the `Model` typealias so that we don't need to worry
about the face that we're working with WorkoutSets externally from this
component.