Simplify single set model
This was overly complicated. It was fun at the beginning, but especially once
we want to add anything other than set information to this model (like the
number of movements, which we need for validation), this gets out of control
quickly. I think it's probably best to simplify this now so that we don't end
up with something over-architected in the long run.