Track movement count in the single set model
We're going to need this for validation purposes. Without it we won't be able
to tell if the rep pattern is valid for the list of movements.

The most straightforward thing to do is to pass the count in when creating the
initial model. This will let us create new models easily with a pre-defined
count.