Add a function for getting the length of a scheme
This is a fairly simple function, but it's nice to have around since it seems
like a common pattern for us. Given a string-based scheme (rep or movement),
this will tell us how many items are in it, by parsing the scheme and then
returning the number of items in it.