Delete Haskell source
I don't think I have the time/energy to use Haskell for this project. This is
super depressing, but since the server-side app is almost entirely going to be
an API for dealing with what is a fairly complex relational object graph, I'm
not super interested in spending all of my time trying to learn how esqueleto
works. Maybe I'll revisit this at some point in the future.