Serve Elm directly from Phoenix
This adds the Elm part of the app into the normal phoenix server process. It
compiles the elm source and adds the compiled JS into assets/vendor (which
we're ignoring from git). This setup will let us automatically build our elm
components from the server process, and easily embed the result in our HTML.

For now we're just embedding the entirety of the current elm app on the main
screen because who cares.