Extract weight selector into standalone component
I moved away from modeling this selection thing as working on the Weight type
specifically because I felt like my model was nested too deeply inside
SingleSet. As it turns out, I think that I was _actually_ just modeling my
components too shallow, and should have been maintaining a 1:1 relationship
between my models and my components. This is the first step towards trying to
make that work, and moves Weight into a standalone component (and isolates the
type itself into a new module) so that it is easier to deal with.