Use a sum type for Weight
I'm not positive about this change, but it was a fun exercise. The general
idea is to work with Weight types as a sum type instead of as a product type.
This resembles how I'd like them to _actually_ look a bit more, but as you can
see it introduces a _lot_ more overhead to the type to make up for the fact
that the units are now constructors and so can't be passed around in quite the
same way.