Use concrete instances for availableWeightUnits
Instead of trying adding a third place to have our String versions of Units,
we can map our existing `unitToString` function over already-constructed
instances of Weight. This means we now only have 2 places where we're using
string values for units, and they are specifically for when moving to or from
strings.