Standardize naming to expect qualification
Naming all of our functions to include the word "weight" was both done
inconsistently, and over-definsively. We should instead expect that we'll be
using the elm-standard (afaict) practice of using a qualified import and
referring to the functions with their module as a namespace. This means that
_internally_ we get to use nice short names, and at the callsite it's still
unambiguous as to which function we're calling.