Add lenses for Weight.amount and Weight.unit
This adds a _super_ simple `Lens` type to the project and uses that to
construct lenses for getting/setting the amount and unit for a weight.

Unfortunately, I'm not sure how to use these with a qualified import. I'd
rather have these named as `amount` and `unit` so that I can do like
`Weight.unit.set` externally, but when I do that it tries to access the module
named `Weight.unit`, which isn't what I want.