Add simple movement input component
This component is super basic right now, but will be the main interface for
adding movements to a movement group (still not sure what to call this). It
basically replaces the existing behavior of using a single text field for
multiple movements. It's likely that this will be broken down into
sub-components eventually.