Track movement index
Previously, when a movement changed, we were just returning that movement
again. That's not actually what we want, though. Instead, in order to handle
changes properly, we need to be able to track which movement changed and edit
it directly.

By tracking the index, we're also able to do things like insert a new set in
the middle of the list by typing `+`.