Update focus when we add a new set
When we're adding a new set, it's always via user interaction, and so we
always want to update the focus into the new field. This is true if we're
hitting the `+` button, or if we've typed "+" to add a new field.

Updating the focus is done through the `Dom` package. The task that it uses to
update the focus can fail, so we need to wrap it in our own function. We don't
really care if the focus fails right now, so we'll just swallow the error.

To make this work, we also need to add a new message, which we'll treat as a
no-op.