Rename FocusedField to NoOp
This is closer to the standard practice, and is _actually_ what we want. We
don't really care about knowing that the field was focused (yet?), so there's
not a whole lot of point in having an explicitly named message for that.
Instead, we can use a basic NoOp message to indicate that we're ignoring the
message entirely.