I was wrong, re-add precondition to onDelete
Nope that was incorrect. This definitely does not get sent only when the field
is empty. So we definitely need to do some checking to ensure that our string
is empty before deleting.

That being said, the way were were handling this before was not great. This
feels a _bit_ better, but still not 100% right. Instead of taking the string,
we'll just take a boolean to use as a precondition. That way, we can offload
the responsibility of checking the string value to the component.

There's still almost certainly a better way to accomplish this, but I can't
see it yet.