Replace SingleSet with WorkoutSet
This removes the old SingleSet version of this in favor of the new WorkoutSet
component. The resulting code is significantly cleaner IMO, and much less
error-prone because we no longer need to do validation and can instead prevent
input errors from happening in the first place.

This is also letting us remove the now-outdated List helpers and the unused
scheme file/parser.