Re-add Main as simple wrapper component
We'll keep using main as our entry point for development purposes, but we
don't want to keep adding things in here. By making this a super simple
component that just wraps another one in a `main` function, we'll be able to
continue to use it in that way without needing to make significant changes
when we want to point it at a new top-level component.